seperated search bar logic
