+++ source/js/buyers-guide/template-js-handler/mobile-search-bar.js
@@ -1,3 +1,5 @@
+import setupSearchBar from "./search-bar";
+
 export default () => {
   const mobileSearch = document.querySelector("#mobile-search");
   const searchContainer = document.querySelector("#pni-mobile-container");
@@ -15,46 +17,15 @@ export default () => {
     searchContainer.classList.toggle("tw-hidden");
   });
 
-  /**
-   * if the user is not on the product review pages,
-   * the search bar has a different behavior that redirects to product review
-   */
-
   if (
     !location.pathname.includes("categories") &&
     !location.pathname.endsWith("/privacynotincluded/")
   ) {
-    const input = document.querySelector("#pni-mobile-bar");
-    const clearIcon = document.querySelector(
-      "#pni-mobile-container .clear-icon"
+    setupSearchBar(
+      "#pni-mobile-bar",
+      "#pni-mobile-container .clear-icon",
+      `#pni-mobile-container`
     );
-
-    clearIcon.addEventListener("click", function () {
-      searchContainer.classList.remove(`has-content`);
-      input.value = "";
-    });
-
-    input.addEventListener(`input`, function () {
-      const searchText = input.value.trim();
-
-      if (searchText) {
-        searchContainer.classList.add(`has-content`);
-      } else {
-        searchContainer.classList.remove(`has-content`);
-      }
-    });
-
-    input.addEventListener("keypress", function (event) {
-      // If the user presses the "Enter" key on the keyboard or mobile
-      if (event.key === "Enter" && input.value) {
-        event.preventDefault();
-        const url = new URL("/privacynotincluded/", location.href);
-        url.searchParams.set("search", input.value);
-        url.search = url.searchParams.toString();
-        url.hash = "product-review";
-        location.href = url.toString();
-      }
-    });
   } else {
     const input = document.querySelector("#pni-mobile-bar");
